Made the adapter functionality more configurable with an automatic fall back if nothing is configured
Moved each of the adapter classes to their own files in the sub-module adapter
