+++ mxcube3/core/adapter/data_publisher_adapter.py
@@ -1,0 +1,33 @@
+from mxcubecore.BaseHardwareObjects import HardwareObjectState
+
+from mxcube3.core.adapter.adapter_base import AdapterBase
+from mxcube3 import server
+
+
+class DataPublisherAdapter(AdapterBase):
+    def __init__(self, ho, name, **kwargs):
+        """
+        Args:
+            (object): Hardware object.
+            (str): The name of the object.
+        """
+        super(DataPublisherAdapter, self).__init__(ho, name, **kwargs)
+
+        try:
+            ho.connect("data", self._new_data_handler)
+            ho.connect("start", self._update_publisher_handler)
+            ho.connect("end", self._update_publisher_handler)
+        except BaseException:
+            msg = "Could not initialize DataPublisherAdapter"
+            logging.getLogger("MX3.HWR").exception(msg)
+        else:
+            self._available = True
+
+    def _new_data_handler(self, data):
+        server.emit("data_publisher_new_data", data, namespace="/hwr")
+
+    def _update_publisher_handler(self, data):
+        server.emit("data_publisher_update", data, namespace="/hwr")
+
+    def state(self):
+        return HardwareObjectState.READY.value
