+++ mxcube3/core/adapter/machine_info_adapter.py
@@ -1,0 +1,75 @@
+from mxcubecore.BaseHardwareObjects import HardwareObjectState
+
+from mxcube3.core.adapter.adapter_base import ActuatorAdapterBase
+from mxcube3.core import utils
+
+from mxcube3.core.models import HOModel, HOMachineInfoModel
+
+
+class MachineInfoAdapter(ActuatorAdapterBase):
+    def __init__(self, ho, name, **kwargs):
+        """
+        Args:
+            (object): Hardware object.
+            (str): The name of the object.
+        """
+        super(MachineInfoAdapter, self).__init__(ho, name, **kwargs)
+        ho.connect("valueChanged", self._value_change)
+
+    def _set_value(self, value):
+        pass
+
+    @utils.RateLimited(0.1)
+    def _value_change(self, *args, **kwargs):
+        self.value_change(self.get_value(), **kwargs)
+
+    def _get_value(self):
+        return {
+            "current": self.get_current(),
+            "message": self.get_message(),
+            "fillmode": self.get_fill_mode(),
+        }
+
+    def get_message(self):
+        try:
+            message = self._ho.getMessage()
+        except (TypeError, AttributeError):
+            message = ""
+
+        return message
+
+    def get_current(self):
+        try:
+            current = self._ho.getCurrent()
+            current = (
+                current
+                if isinstance(current, str)
+                else "{:.1f} mA".format(round(float(self._ho.getCurrent()), 1))
+            )
+        except (TypeError, AttributeError):
+            current = "-1"
+
+        return current
+
+    def get_fill_mode(self):
+        try:
+            fmode = self._ho.getFillMode()
+        except (TypeError, AttributeError):
+            fmode = ""
+
+        return fmode
+
+    def limits(self):
+        """
+        Returns: The detector distance limits.
+        """
+        return (-1, -1)
+
+    def stop(self):
+        pass
+
+    def state(self):
+        return HardwareObjectState.READY.value
+
+    def dict(self):
+        return HOMachineInfoModel(**self._dict_repr()).dict()
