+++ mxcube3/core/beamlineutils.py
@@ -12,7 +12,7 @@ from mxcube3 import mxcube
 from .qutils import READY
 
 
-from mxcube3.core.beamline_adapter import BeamlineAdapter
+from mxcube3.core.adapter.beamline_adapter import BeamlineAdapter
 
 
 def init_signals():
@@ -274,24 +274,24 @@ def beamline_run_action(name, params):
         raise Exception(msg)
 
 
-def beamline_set_attribute(name, data):
-    """
-    """
-    ho = mxcube.mxcubecore.beamline.get_object(name.lower())
-    data = ho.set_value(data["value"])
-    res = data["available"]     
+# def beamline_set_attribute(name, data):
+#     """
+#     """
+#     ho = mxcube.mxcubecore.beamline.get_object(name.lower())
+#     data = ho.set_value(data["value"])
+#     res = data["available"]     
 
-    return res, data
+#     return res, data
 
 
-def beamline_get_attribute(name):
-    """
-    """
-    ho = mxcube.mxcubecore.beamline.get_object(name.lower())
-    data = ho.dict()
-    res = 200 if data["available"] else 520
+# def beamline_get_attribute(name):
+#     """
+#     """
+#     ho = mxcube.mxcubecore.beamline.get_object(name.lower())
+#     data = ho.dict()
+#     res = 200 if data["available"] else 520
 
-    return res, data
+#     return res, data
 
 
 def get_beam_info():
