+++ mxcube3/core/qutils.py
@@ -24,7 +24,6 @@ from mxcube3 import server
 from . import limsutils
 from . import utils
 
-from .beamline_adapter import BeamlineAdapter
 from functools import reduce
 
 # Important: same constants as in constants.js
